+++ src/automatica.drivers/automatica.driver.homekit/P3.Driver.HomeKit/HomeKitServer.cs
@@ -147,7 +147,7 @@ namespace P3.Driver.HomeKit
             return true;
         }
 
-        public int AddAccessory(Accessory accessory)
+        public ulong AddAccessory(Accessory accessory)
         {
             _accessoryContainer.AddAccessory(accessory);
             return accessory.Id;
